% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_forest.R
\name{delete_forest}
\alias{delete_forest}
\title{Clean up Dirichlet Forest}
\usage{
delete_forest(forest_model)
}
\arguments{
\item{forest_model}{A trained forest model}
}
\description{
Free memory used by the forest model
}
