% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_forest.R
\name{predict_dirichlet_forest}
\alias{predict_dirichlet_forest}
\title{Predict with Dirichlet Forest}
\usage{
predict_dirichlet_forest(forest_model, X_new)
}
\arguments{
\item{forest_model}{A trained forest model from dirichlet_forest()}

\item{X_new}{Numeric matrix of new predictors}
}
\value{
A list with alpha_predictions and mean_predictions matrices
}
\description{
Make predictions using a trained Dirichlet Forest
}
